INT-1229 make the community tab stateless
