--- intuita-webview/src/communityView/index.html
@@ -1,13 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-	<head>
-		<meta charset="utf-8" />
-		<meta name="viewport" content="width=device-width, initial-scale=1" />
-		<meta name="theme-color" content="#000000" />
-		<title>Intuita</title>
-	</head>
-	<body>
-		<div id="root"></div>
-		<script type="module" src="./index.tsx"></script>
-	</body>
-</html>
