--- intuita-webview/src/communityView/index.tsx
@@ -1,14 +1,0 @@
-import React from 'react';
-import ReactDOM from 'react-dom/client';
-import '../shared/index.css';
-import App from './App';
-
-const root = ReactDOM.createRoot(
-	document.getElementById('root') as HTMLElement,
-);
-
-root.render(
-	<React.StrictMode>
-		<App screenWidth={null} />
-	</React.StrictMode>,
-);
