+++ src/components/webview/MainProvider.ts
@@ -118,7 +118,6 @@ export class MainViewProvider implements WebviewViewProvider {
 		const codemodRunsProps = this.__codemodRuns.getInitialProps();
 		const fileExplorerProps = this.__fileExplorer.getInitialProps();
 		const codemodListProps = this.__codemodList.getInitialProps();
-		const communityProps = this.__community.getInitialProps();
 
 		this.__webviewResolver.resolveWebview(
 			webviewView.webview,
@@ -127,7 +126,6 @@ export class MainViewProvider implements WebviewViewProvider {
 				codemodRunsProps,
 				fileExplorerProps,
 				codemodListProps,
-				communityProps,
 			}),
 		);
 	}
